+++ src/elusion/zenopay/config.py
@@ -26,6 +26,7 @@ DEFAULT_HEADERS = {
 # API endpoints
 ENDPOINTS = {
     "create_order": "/api/payments/mobile_money_tanzania",
+    "checkout": "/api/payments/checkout/",
     "order_status": "/api/payments/order-status",
     "disbursement": "/api/payments/walletcashin/process/",
     "utility-payments": "/api/payments/utilitypayment/process/",